Supply now drops in doubles

+++ src/main/battlecode/world/GameMap.java
@@ -247,17 +247,17 @@ public class GameMap implements GenericGameMap {
     public static class MapMemory {
         private final GameMap map;
         private final boolean[][] seen;
-        private final int[][] supplyLevel;
+        private final double[][] supplyLevel;
         private final double[][] oreMined;
 
         public MapMemory(GameMap map) {
             this.map = map;
             this.seen = new boolean[map.getWidth()][map.getHeight()];
-            this.supplyLevel = new int[map.getWidth()][map.getHeight()];
+            this.supplyLevel = new double[map.getWidth()][map.getHeight()];
             this.oreMined = new double[map.getWidth()][map.getHeight()];
         }
 
-        public void rememberLocations(MapLocation loc, int radiusSquared, Map<MapLocation, Integer> droppedSupplies, Map<MapLocation, Double> oreMinedMap) {
+        public void rememberLocations(MapLocation loc, int radiusSquared, Map<MapLocation, Double> droppedSupplies, Map<MapLocation, Double> oreMinedMap) {
             MapLocation[] locs = MapLocation.getAllMapLocationsWithinRadiusSq(loc, radiusSquared);
 
             for (int i = 0; i < locs.length; i++) {
@@ -299,7 +299,7 @@ public class GameMap implements GenericGameMap {
             }
         }
 
-        public int recallSupplyLevel(MapLocation loc) {
+        public Double recallSupplyLevel(MapLocation loc) {
             int X = loc.x - map.mapOriginX;
             int Y = loc.y - map.mapOriginY;
